+++ pyjinhx/builtins/ui/pjx_modal_body/pjx_modal_body.py
@@ -2,8 +2,6 @@ from pyjinhx import BaseComponent
 from pyjinhx.base import AttrValue
 
 
-class PJXModal(BaseComponent):
-    open_on_mount: bool = False
-    remove_on_close: bool = False
+class PJXModalBody(BaseComponent):
     class_name: AttrValue = ""
     content: str | BaseComponent = ""